Updated the project structure
Updated the project structure - files are split correctly into folders
